feat: Setup Robust CI/CD Pipeline (#11)
* feat: setup robust ci/cd pipeline

BREAKING CHANGE: This commit introduces a new CI/CD pipeline structure, replacing the old single-file workflow. It splits the process into two separate workflows: 'ci.yml' for continuous integration (linting, testing, security scans, and building) and 'release.yml' for handling releases with GoReleaser on tag creation. The new 'ci.yml' is triggered on pull requests and pushes to the master branch, enhancing code quality and validation. The 'release.yml' ensures that releases are generated only when a new version tag is pushed, fixing the previous issue of releases being triggered on every push to master. This change requires developers to adapt to the new workflow structure and trigger conditions.

* feat(ci): update ci and release workflows

- Update Go version to 1.25.1
- Add Git installation step
- Set GOPROXY to direct
- Update goreleaser-action to v2

* fix(ci): correct CI workflow configuration

* fix(ci): add path to upload-artifact action

* feat(deps): vendor dependencies

---------

Co-authored-by: Igor Guedes Rodrigues <chalkan3@MacBook-Air-de-Igor.local>